% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth_NLU.R
\name{auth_NLU}
\alias{auth_NLU}
\title{Watson Natural Language Understanding API Authentication}
\usage{
auth_NLU(username = NULL, password = NULL)
}
\arguments{
\item{username}{Authentication IBM Watson Natural-Language-Understanding-3j \strong{username}}

\item{password}{Authentication IBM Watson Natural-Language-Understanding-3j \strong{password}}
}
\value{
If authentication is successful, there is no return value. If unsuccessful,
   the function will ask the user to ensure username and password combination are correct.
}
\description{
See the \href{https://github.com/johannesharmse/watsonNLU/blob/master/README.md}{sign-up} documentation for step by step instructions to secure your own username and password to enable you to use the Watson NLU API.
}
\examples{
# Find username and password under service credentials
auth_NLU(username = 'XXXXXXXX-XXXX-XXXX-XXXX-XXXXXXXXXXXX',
 password= 'XXXXXXXXXXXX')

}
