% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyword_relevance.R
\name{keyword_relevance}
\alias{keyword_relevance}
\title{Watson Natural Language Understanding: Relevance of Keywords}
\usage{
keyword_relevance(text = NULL, url = NULL)
}
\arguments{
\item{text}{text string to be categorized.
Either \code{text} or \code{url} argument has to be specified,
but not both.}

\item{url}{url to text to be categorized.
Either \code{text} or \code{url} argument has to be specified,
but not both.}
}
\value{
A dataframe that contains keywords and their likelihood of being a keyword, for the given input.
}
\description{
The \strong{keyword_relevance} function takes in a URL or text input, and returns a dataframe that contains keywords and their likelihood of being a keyword, from the given input.
}
\details{
See the \href{https://github.com/johannesharmse/watsonNLU/blob/master/README.md}{keyword_relevance} documentation for more useage cases.
}
\examples{
# Find the likelihood that a word is a keyword from the given input: text input
keyword_relevance(text = 'This is a great API wrapper')

# Find the likelihood that a word is a keyword from the given input: URL input
keyword_relevance(url = 'http://santiago.begueria.es/2010/10/generating-spatially-correlated-random-fields-with-r/')

}
