% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyword_sentiment.R
\name{keyword_sentiment}
\alias{keyword_sentiment}
\title{Watson Natural Language Understanding: Sentiment Analysis of Input Keywords}
\usage{
keyword_sentiment(input = NULL, input_type = NULL,
  version = "?version=2018-03-16")
}
\arguments{
\item{input}{Either a text string input or website URL.
Either \code{text} or \code{url} argument has to be specified,
but not both.}

\item{input_type}{Specify what type of input was entered.
Either \code{text} or \code{url} argument has to be specified,
but not both.}

\item{version}{The release date of the API version to use. Default value is \code{version="?version=2018-03-16"}}
}
\value{
Dataframe containing the sentiments of the keywords extracted from the input, and the likelihood that the input is described by that sentiment.
}
\description{
See the \href{https://github.com/johannesharmse/watsonNLU/blob/master/README.md}{sign-up} documentation for step by step instructions to secure your own username and password to enable you to authenticate the Watson NLU API. The \strong{keyword_sentiment} function takes a text or URL input, along with the input type. The function then returns a dataframe containing the sentiments of the keywords extracted from the input, and the likelihood that the input is described by that sentiment. See the \href{https://github.com/johannesharmse/watsonNLU/blob/master/README.md}{keyword_sentiment} documentation for more useage cases.
}
\examples{
library(watsonNLU)

# Authenticate using Watson NLU API Credentials
auth_NLU(username="XXXX", password="XXXX")

# Find the keywords and related emotions in the given text input.
keyword_sentiment(input = 'This is a great API wrapper', input_type='text')

# Find the keywords and related emotions in the given URL input.
keyword_sentiment(input = 'http://www.nytimes.com/guides/well/how-to-be-happy', input_type='url')

}
\seealso{
\code{\link[watsonNLU]{text_categories}}, \code{\\link[watsonNLU]{keyword_relevance}}, \code{\\link[watsonNLU]{keyword_emotions}}, \code{\\link[watsonNLU]{auth_NLU}}
}
