\name{WatsonNLU-package}
\alias{WatsonNLU-package}
\alias{WatsonNLU}
\docType{package}
\title{
\packageTitle{WatsonNLU}
}
\description{
\packageDescription{WatsonNLU}
}
\details{
}
\author{
\packageAuthor{WatsonNLU}

Maintainer: \packageMaintainer{WatsonNLU}
}
\references{
https://www.ibm.com/watson/developercloud/natural-language-understanding/api/v1/
}
\keyword{ package }
\seealso{
\link[WatsonNLU]{Watson_NLU}
}
\examples{
}
