% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth_NLU.R
\name{auth_NLU}
\alias{auth_NLU}
\title{Watson Natural Language Understanding API Authentication}
\usage{
auth_NLU(username = NULL, password = NULL)
}
\arguments{
\item{username}{Authentication IBM Watson Natural-Language-Understanding-3j \strong{username}}

\item{password}{Authentication IBM Watson Natural-Language-Understanding-3j \strong{password}}
}
\value{
If authentication is successful, there is no return value. If unsuccessful,
   the function will ask the user to ensure username and password combination are correct.
}
\description{
See the \href{https://github.com/johannesharmse/watsonNLU/blob/master/README.md}{sign-up} documentation for step by step instructions to secure your own username and password to enable you to authenticate the Watson NLU API. The \strong{auth_NLU} function takes in a username and password as input to authenticate the users computer to use the Watson Natural Language Understanding API.
}
\examples{
library(watsonNLU)

# Authenticate using Watson NLU API Credentials
auth_NLU(username="XXXX", password="XXXX")

}
\seealso{
\code{\link[watsonNLU]{keyword_sentiment}}, \code{\\link[watsonNLU]{keyword_relevance}}, \code{\\link[watsonNLU]{keyword_emotions}}, \code{\\link[watsonNLU]{text_categories}}
}
