% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyword_sentiment.R
\name{keyword_sentiment}
\alias{keyword_sentiment}
\title{Watson Natural Language Understanding: Document Keyword-Sentiments Analyzer}
\usage{
keyword_sentiment(input = NULL, input_type = NULL,
  version = "?version=2018-03-16")
}
\arguments{
\item{input}{A text input or URL of a website}

\item{input_type}{The type of input. It can either be `text` or `url`, but not both.}

\item{version}{The release date of the API version to use.}
}
\value{
A dataframe with keywords and likelihood of emotions related to those keywords found in the text or URL.

# @examples

# Find the keywords and related emotions in the given text input. By default it takes version = 2018-03-16
keyword_sentiment(input = 'This is a great API wrapper', input_type='text')

# Find the keywords and related emotions in the given URL input. By default it takes version = 2018-03-16
keyword_sentiment(input = 'http://www.nytimes.com/guides/well/how-to-be-happy', input_type='url')
}
\description{
See the \href{https://www.ibm.com/watson/developercloud/natural-language-understanding/api/v1/#get-analyze}{IBM Watson NLU API} documentation for more information.
}
