% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_categories.R
\name{text_categories}
\alias{text_categories}
\title{Watson Natural Language Understanding: Text Categorizer}
\usage{
text_categories(text = NULL, url = NULL)
}
\arguments{
\item{text}{text string to be categorized.
Either \code{text} or \code{url} argument has to be specified,
but not both.}

\item{url}{url to text to be categorized.
Either \code{text} or \code{url} argument has to be specified,
but not both.}
}
\value{
A dataframe that contains the likelihood that the contents of the URL or text belong to a particular category.
}
\description{
The \strong{text_categories} function takes in a URL or text input, and returns a dataframe that contains the likelihood that the contents of the URL or text belong to a particular category.
}
\details{
See the \href{https://github.com/johannesharmse/watsonNLU/blob/master/README.md}{text_categories} documentation for more useage cases.
}
\examples{
# Find the likelihood that the given input belongs to a particular category: from text
text_categories(text = 'This is a great API wrapper')

# Find the likelihood that the given input belongs to a particular category: from URL
text_categories(url = 'http://santiago.begueria.es/2010/10/generating-spatially-correlated-random-fields-with-r/')

}
