% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_categories.R
\name{text_categories}
\alias{text_categories}
\title{Watson Natural Language Understanding: Text Categorizer}
\usage{
text_categories(input = NULL, input_type = NULL, limit = NULL,
  version = "?version=2018-03-16")
}
\arguments{
\item{input}{Either a text string input or website URL.
Either \code{text} or \code{url} argument has to be specified,
but not both.}

\item{input_type}{Specify what type of input was entered.
Either \code{text} or \code{url} argument has to be specified,
but not both.}

\item{limit}{The number of categories to return.}

\item{version}{The release date of the API version to use. Default value is \code{version="?version=2018-03-16"}}
}
\value{
A dataframe that contains the likelihood that the contents of the URL or text belong to a particular category.
}
\description{
See the \href{https://github.com/johannesharmse/watsonNLU/blob/master/README.md}{sign-up} documentation for step by step instructions to secure your own username and password to enable you to use the Watson NLU API. The \strong{text_categories} function takes a text or URL input along with the input type. The function then returns a dataframe that contains the likelihood that the contents of the URL or text belong to a particular category. See the \href{https://github.com/johannesharmse/watsonNLU/blob/master/README.md}{text_categories} documentation for more useage cases.
}
\examples{
library(watsonNLU)

# Authenticate using Watson NLU API Credentials
auth_NLU(username="XXXX", password="XXXX")

# Find 5 categories that describe the text input.
text_categories(input = 'This is a great API wrapper', input_type='text', limit = 5)

# Find 5 categories that describe the URL input.
text_categories(input = 'http://www.nytimes.com/guides/well/how-to-be-happy', input_type='url', limit = 5)

}
